refactor: organize components by type
