+++ .vscode/settings.json
@@ -5,6 +5,7 @@
     "FULLNAME",
     "iife",
     "listbox",
+    "navs",
     "networkidle",
     "onspringend",
     "signin",
