+++ componentObjectModels/dialogs/addLayoutItemDialog.ts
