+++ componentObjectModels/dialogs/baseDeleteDialog.ts
