+++ componentObjectModels/dialogs/deleteAppDialog.ts
