+++ componentObjectModels/modals/editAppNotesSettingsModal.ts
