+++ pageObjectModels/appAdminPage.ts
@@ -1,7 +1,7 @@
 import { Locator, Page } from '@playwright/test';
-import { LayoutItemType } from '../componentObjectModels/addLayoutItemMenu';
-import { AppGeneralTab } from '../componentObjectModels/appGeneralTab';
-import { AppLayoutTab } from '../componentObjectModels/appLayoutTab';
+import { LayoutItemType } from '../componentObjectModels/menus/addLayoutItemMenu';
+import { AppGeneralTab } from '../componentObjectModels/tabs/appGeneralTab';
+import { AppLayoutTab } from '../componentObjectModels/tabs/appLayoutTab';
 import { BASE_URL } from '../playwright.config';
 import { BaseAdminPage } from './baseAdminPage';
