+++ pageObjectModels/appsAdminPage.ts
@@ -1,7 +1,7 @@
 import { Locator, Page } from '@playwright/test';
-import { CreateAppDialog } from '../componentObjectModels/createAppDialog';
-import { CreateAppModal } from '../componentObjectModels/createAppModal';
-import { DeleteAppDialog } from '../componentObjectModels/deleteAppDialog';
+import { CreateAppDialog } from '../componentObjectModels/dialogs/createAppDialog';
+import { DeleteAppDialog } from '../componentObjectModels/dialogs/deleteAppDialog';
+import { CreateAppModal } from '../componentObjectModels/modals/createAppModal';
 import { BaseAdminPage } from './baseAdminPage';
 
 export class AppsAdminPage extends BaseAdminPage {
