+++ pageObjectModels/groupsSecurityAdminPage.ts
@@ -1,5 +1,5 @@
 import { Locator, Page } from '@playwright/test';
-import { DeleteGroupDialog } from '../componentObjectModels/deleteGroupDialog';
+import { DeleteGroupDialog } from '../componentObjectModels/dialogs/deleteGroupDialog';
 import { BaseAdminPage } from './baseAdminPage';
 
 export class GroupsSecurityAdminPage extends BaseAdminPage {
