+++ pageObjectModels/rolesSecurityAdminPage.ts
@@ -1,5 +1,5 @@
 import { Locator, Page } from '@playwright/test';
-import { DeleteRoleDialog } from '../componentObjectModels/deleteRoleDialog';
+import { DeleteRoleDialog } from '../componentObjectModels/dialogs/deleteRoleDialog';
 import { BaseAdminPage } from './baseAdminPage';
 
 export class RolesSecurityAdminPage extends BaseAdminPage {
